Refactor privacy protection components and update related imports
