+++ packages/design-system/src/icons/index.tsx
@@ -45,8 +45,8 @@ export { default as PrivateAdvertisingIcon } from './private-advertising.svg';
 export { default as PrivateAdvertisingIconWhite } from './private-advertising-white.svg';
 export { default as RelatedWebsiteSetsIcon } from './related-website-sets.svg';
 export { default as RelatedWebsiteSetsIconWhite } from './related-website-sets-white.svg';
-export { default as AntiCovertTrackingIcon } from './anti-covert-tracking.svg';
-export { default as AntiCovertTrackingIconWhite } from './anti-covert-tracking-white.svg';
+export { default as PrivacyProtectionIcon } from './anti-covert-tracking.svg';
+export { default as PrivacyProtectionIconWhite } from './anti-covert-tracking-white.svg';
 export { default as InspectIcon } from './inspect.svg';
 export { default as InspectActiveIcon } from './inspect-active.svg';
 export { default as InspectWhiteIcon } from './inspect-white.svg';
