+++ packages/extension/src/view/devtools/pages/privateAdvertising/attributionReporting/index.tsx
@@ -20,8 +20,10 @@ import React, { useMemo } from 'react';
 import {
   PSInfoKey,
   TabsProvider,
+  LandingPage,
   type TabItems,
 } from '@google-psat/design-system';
+import { I18n } from '@google-psat/i18n';
 
 /**
  * Internal dependencies.
@@ -70,10 +72,18 @@ const AttributionReporting = () => {
     []
   );
 
-  return (
+  const contentPanel = (
     <TabsProvider items={tabItems} name="attribution-reporting">
       <Panel />
     </TabsProvider>
+  );
+
+  return (
+    <LandingPage
+      title={I18n.getMessage('attributionReporting')}
+      contentPanel={contentPanel}
+      hasTabs={true}
+    />
   );
 };
