+++ packages/extension/src/view/devtools/pages/privateAdvertising/protectedAudience/index.tsx
@@ -20,8 +20,10 @@ import React, { useMemo } from 'react';
 import {
   PSInfoKey,
   TabsProvider,
+  LandingPage,
   type TabItems,
 } from '@google-psat/design-system';
+import { I18n } from '@google-psat/i18n';
 
 /**
  * Internal dependencies.
@@ -93,10 +95,18 @@ const ProtectedAudience = () => {
     []
   );
 
-  return (
+  const contentPanel = (
     <TabsProvider items={tabItems} name="protectedAudience">
       <Panel />
     </TabsProvider>
+  );
+
+  return (
+    <LandingPage
+      contentPanel={contentPanel}
+      title={I18n.getMessage('protectedAudience')}
+      hasTabs={true}
+    />
   );
 };
